02.Student_Performance_ML_E2E_project/requirements.txt
numpy
pandas==1.1.3
matplotlib
seaborn
scikit-learn>=0.24.0
flask
mysql-connector-python
pymysql

python-dotenv  # This helps in reading the envirnoment variable.... .env

# -e .
# ipykernel we will install in the local...using"  pip install ipykernel "
